% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.catboost-package}
\alias{mlr3learners.catboost}
\alias{mlr3learners.catboost-package}
\title{mlr3learners.catboost: Learners from catboost package for mlr3}
\description{
Adds Learner functionality from the catboost
    package to mlr3.
}
\author{
\strong{Maintainer}: Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})

}
