% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrCatboost.R
\name{mlr_learners_regr.catboost}
\alias{mlr_learners_regr.catboost}
\alias{LearnerRegrCatboost}
\title{Gradient Boosted Decision Trees Regression Learner}
\description{
Gradient boosted decision trees regressionlearner.
Uses \code{\link[catboost:catboost.train]{catboost::catboost.train()}} and \code{\link[catboost:catboost.predict]{catboost::catboost.predict()}} from
package catboost.

By default, the RMSE is optimized as the loss function.
}
\section{Custom mlr3 defaults}{

\itemize{
\item \code{logging_level}:
\itemize{
\item Actual default: "Verbose"
\item Adjusted default: "Silent"
\item Reason for change: consistent with other mlr3 learners
}
\item \code{thread_count}:
\itemize{
\item Actual default: -1
\item Adjusted default: 1
\item Reason for change: consistent with other mlr3 learners
}
\item \code{allow_writing_files}:
\itemize{
\item Actual default: TRUE
\item Adjusted default: FALSE
\item Reason for change: consistent with other mlr3 learners
}
\item \code{save_snapshot}:
\itemize{
\item Actual default: TRUE
\item Adjusted default: FALSE
\item Reason for change: consistent with other mlr3 learners
}
}
}

\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("regr.catboost")
lrn("regr.catboost")
}
}

\examples{
if (requireNamespace("catboost")) {
  learner = mlr3::lrn("regr.catboost")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
}
\references{
\cite{mlr3learners.catboost}{dorogush2019}
\cite{mlr3learners.catboost}{prokhorenkova2017}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}:
\link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrCatboost}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerRegrCatboost$new()}}
\item \href{#method-importance}{\code{LearnerRegrCatboost$importance()}}
\item \href{#method-clone}{\code{LearnerRegrCatboost$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{LearnerRegrCatboost} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrCatboost$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importance"></a>}}
\if{latex}{\out{\hypertarget{method-importance}{}}}
\subsection{Method \code{importance()}}{
The importance scores are calculated using
\code{\link[catboost:catboost.get_feature_importance]{catboost.get_feature_importance}},
setting \code{type = "FeatureImportance"}, returned for 'all'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrCatboost$importance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named \code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrCatboost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
